requirements.txt
groq
python-dotenv
uvicorn[standard]
pypdf
chromadb
sentence-transformers